@@ -24,10 +24,9 @@
     "@types/typescript": "^2.0.0",
     "typedoc": "^0.13.0",
     "typedoc-plugin-markdown": "^1.1.18",
-    "typescript": "2.6.2"
+    "typescript": "^3.8.3"
   },
   "dependencies": {
-    "request": "^2.85.0",
     "string-template": "^1.0.0"
   }
 }
