@@ -12,9 +12,12 @@ Object.defineProperty(exports, "__esModule", { value: true });
 const param = require("./parameters");
 const Quota_1 = require("./Quota");
 const fs = require("fs");
+const http = require("http");
+const https = require("https");
 const os = require("os");
 const process = require("process");
-const request = require("request");
+const querystring = require("querystring");
+const url = require("url");
 const format = require("string-template");
 const customErrors_1 = require("./customErrors");
 function handleRestResult(resolve, reject, url, error, response, body, onUpdateLimit) {
@@ -63,52 +66,78 @@ function handleRestResult(resolve, reject, url, error, response, body, onUpdateL
         reject(new Error(`failed to parse server response for request "${url}": ${err.message}`));
     }
 }
-function restGet(url, args, onUpdateLimit) {
-    const stackErr = new Error();
+function lib(inputUrl) {
+    return url.parse(inputUrl).protocol === 'http:'
+        ? http
+        : https;
+}
+function restGet(inputUrl, args, onUpdateLimit) {
     return new Promise((resolve, reject) => {
-        const finalURL = format(url, args.path || {});
-        request.get(finalURL, {
-            headers: args.headers,
-            followRedirect: true,
-            timeout: args.requestConfig.timeout,
-            agentOptions: {
-                secureProtocol: args.requestConfig.securityProtocol
-            },
-        }, (error, response, body) => {
-            if (error) {
-                error.message += ` (request: ${url})`;
-                error.stack += '\n' + stackErr.stack;
+        const started = Date.now();
+        const finalURL = format(inputUrl, args.path || {});
+        lib(inputUrl).get(Object.assign(Object.assign({}, url.parse(finalURL)), { headers: args.headers, timeout: args.requestConfig.timeout }), (res) => {
+            const { statusCode } = res;
+            const contentType = res.headers['content-type'];
+            let err;
+            if (statusCode !== 200) {
+                err = `Request Failed. Status Code: ${statusCode}`;
             }
-            handleRestResult(resolve, reject, url, error, response, body, onUpdateLimit);
+            else if (!/^application\/json/.test(contentType)) {
+                err = `Invalid content-type ${contentType}`;
+            }
+            if (err !== undefined) {
+                res.resume();
+                return reject(new Error(err));
+            }
+            res.setEncoding('utf8');
+            let rawData = '';
+            res
+                .on('data', (chunk) => { rawData += chunk; })
+                .on('error', err => {
+                handleRestResult(resolve, reject, inputUrl, err, res, rawData, onUpdateLimit);
+            })
+                .on('end', () => {
+                handleRestResult(resolve, reject, inputUrl, null, res, rawData, onUpdateLimit);
+            });
         });
     });
 }
-function restPost(method, url, args, onUpdateLimit) {
+function restPost(method, inputUrl, args, onUpdateLimit) {
     const stackErr = new Error();
     return new Promise((resolve, reject) => {
-        const finalURL = format(url, args.path);
-        console.log('POST', finalURL);
-        request({
-            method,
-            url: finalURL,
-            headers: args.headers,
-            followRedirect: true,
-            timeout: args.requestConfig.timeout,
-            agentOptions: {
-                secureProtocol: args.requestConfig.securityProtocol
-            },
-            body: JSON.stringify(args.data),
-        }, (error, response, body) => {
-            if (error) {
-                error.message += ` (request: ${url})`;
-                error.stack += '\n' + stackErr.stack;
-            }
-            handleRestResult(resolve, reject, url, error, response, body, onUpdateLimit);
+        const finalURL = format(inputUrl, args.path);
+        const body = JSON.stringify(args.data, undefined, 2);
+        const headers = Object.assign(Object.assign({}, args.headers), { 'Content-Length': body.length });
+        if (process.env.APIKEYMASTER !== undefined) {
+            headers['apikeymaster'] = process.env.APIKEYMASTER;
+        }
+        const req = lib(inputUrl).request(Object.assign(Object.assign({}, url.parse(finalURL)), { method,
+            headers, timeout: args.requestConfig.timeout }), (res) => {
+            res.setEncoding('utf8');
+            let rawData = '';
+            res
+                .on('data', (chunk) => { rawData += chunk; })
+                .on('error', err => {
+                handleRestResult(resolve, reject, inputUrl, err, res, rawData, onUpdateLimit);
+            })
+                .on('end', () => {
+                const { statusCode } = res;
+                const contentType = res.headers['content-type'];
+                let err = null;
+                if (statusCode !== 200) {
+                    err = new customErrors_1.HTTPError(res.statusCode, res.statusMessage, rawData);
+                }
+                else if (!/^application\/json/.test(contentType)) {
+                    err = new Error(`Invalid content-type ${contentType}`);
+                }
+                handleRestResult(resolve, reject, inputUrl, err, res, rawData, onUpdateLimit);
+            });
         });
+        req.write(body);
+        req.end();
     });
 }
 function rest(url, args, onUpdateLimit, method) {
-    console.log('rest request', url, args);
     return args.data !== undefined
         ? restPost(method || 'POST', url, args, onUpdateLimit)
         : restGet(url, args, onUpdateLimit);
@@ -116,6 +145,7 @@ function rest(url, args, onUpdateLimit, method) {
 class Nexus {
     constructor(appName, appVersion, defaultGame, timeout) {
         this.mBaseURL = param.API_URL;
+        this.mGraphBaseURL = param.GRAPHQL_URL;
         this.mRateLimit = { daily: 1000, hourly: 100 };
         this.mBaseData = {
             headers: {
@@ -131,7 +161,6 @@ class Nexus {
                 gameId: defaultGame,
             },
             requestConfig: {
-                securityProtocol: param.SECURITY_PROTOCOL_VERSION,
                 timeout: timeout || param.DEFAULT_TIMEOUT_MS,
                 noDelay: true,
             },
@@ -242,57 +271,6 @@ class Nexus {
             }));
         });
     }
-    getCollectionsByGame(gameId) {
-        return __awaiter(this, void 0, void 0, function* () {
-            yield this.mQuota.wait();
-            const res = yield this.request(this.mBaseURL + '/games/{gameId}/collections', this.args({
-                path: this.filter({ gameId }),
-            }));
-            return res.collections;
-        });
-    }
-    getCollectionsByUser(userId) {
-        return __awaiter(this, void 0, void 0, function* () {
-            yield this.mQuota.wait();
-            const res = yield this.request(this.mBaseURL + '/users/{userId}/collections', this.args({
-                path: this.filter({ userId }),
-            }));
-            return res.collections;
-        });
-    }
-    getRevisions(collectionId) {
-        return __awaiter(this, void 0, void 0, function* () {
-            yield this.mQuota.wait();
-            const res = yield this.request(this.mBaseURL + '/collections/{collectionId}/revisions', this.args({
-                path: this.filter({ collectionId }),
-            }));
-            return res.collection_revisions;
-        });
-    }
-    getRevisionMods(collectionId, revisionId) {
-        return __awaiter(this, void 0, void 0, function* () {
-            yield this.mQuota.wait();
-            return this.request(this.mBaseURL + '/collections/{collectionId}/revisions/{revisionId}/revision_mods', this.args({
-                path: this.filter({ collectionId, revisionId }),
-            }));
-        });
-    }
-    getCollectionImages(collectionId) {
-        return __awaiter(this, void 0, void 0, function* () {
-            yield this.mQuota.wait();
-            return this.request(this.mBaseURL + '/collections/{collectionId}/images', this.args({
-                path: this.filter({ collectionId }),
-            }));
-        });
-    }
-    getCollectionVideos(collectionId) {
-        return __awaiter(this, void 0, void 0, function* () {
-            yield this.mQuota.wait();
-            return this.request(this.mBaseURL + '/collections/{collectionId}/videos', this.args({
-                path: this.filter({ collectionId }),
-            }));
-        });
-    }
     getEndorsements() {
         return __awaiter(this, void 0, void 0, function* () {
             yield this.mQuota.wait();
@@ -397,94 +375,50 @@ class Nexus {
             }
         });
     }
-    sendCollection(manifest, assetFilePath, gameId) {
+    createCollection(data) {
         return __awaiter(this, void 0, void 0, function* () {
             yield this.mQuota.wait();
-            return new Promise((resolve, reject) => {
-                const formData = {
-                    collection_schema_id: 1,
-                    name: manifest.info.name,
-                    description: manifest.info.description,
-                    summary: manifest.info.description,
-                    adult_content: manifest.info.adult_content ? 'true' : 'false',
-                    collection_manifest: JSON.stringify(manifest),
-                    collection_data: fs.createReadStream(assetFilePath),
-                };
-                if (manifest.info.collection_id !== undefined) {
-                    formData['collection_id'] = manifest.info.collection_id;
-                }
-                const baseUrl = param.API_DEV_URL || param.API_URL;
-                const url = `${baseUrl}/games/${gameId || this.mBaseData.path.gameId}/collections/revisions`;
-                const headers = Object.assign({}, this.mBaseData.headers);
-                if (!!param.APIKEY_DEV) {
-                    headers.APIKEY = param.APIKEY_DEV;
-                }
-                request.post({
-                    headers,
-                    url,
-                    formData,
-                    timeout: this.mBaseData.requestConfig.timeout,
-                }, (error, response, body) => {
-                    if (error !== null) {
-                        return reject(error);
-                    }
-                    else if (response.statusCode >= 400) {
-                        if (response.statusCode === 422) {
-                            try {
-                                const parsed = JSON.parse(body);
-                                const ex = parsed.message !== undefined
-                                    ? new customErrors_1.ParameterInvalid(parsed.message)
-                                    : new customErrors_1.ParameterInvalid(parsed);
-                                if (parsed.name !== undefined) {
-                                    ex.name = parsed.name;
-                                }
-                                Object.keys(parsed).forEach(key => {
-                                    if (['name', 'message'].indexOf(key) === -1) {
-                                        ex[key] = parsed[key];
-                                    }
-                                });
-                                return reject(ex);
-                            }
-                            catch (err) {
-                                return reject(new customErrors_1.ParameterInvalid(body));
-                            }
-                        }
-                        else {
-                            return reject(new customErrors_1.HTTPError(response.statusCode, response.statusMessage, body));
-                        }
-                    }
-                    else {
-                        const result = JSON.parse(body);
-                        return resolve(result.collection_revision);
-                    }
-                });
-            });
+            return yield this.mutateGraph('createCollection', {
+                collectionData: { type: 'CollectionPayload', optional: false },
+            }, { collectionData: data }, this.args({ path: this.filter({}) }));
         });
     }
-    getCollectionInfo(collectionId) {
+    updateCollection(data, collectionId) {
         return __awaiter(this, void 0, void 0, function* () {
             yield this.mQuota.wait();
-            const res = yield this.request(this.mBaseURL + '/collections/{collectionId}', this.args({ path: this.filter({ collectionId }) }));
-            return res.collection;
+            return yield this.mutateGraph('updateCollection', {
+                collectionData: { type: 'CollectionPayload', optional: false },
+                collectionId: { type: 'Int', optional: false },
+            }, { collectionData: data, collectionId }, this.args({ path: this.filter({}) }));
         });
     }
-    getRevisionInfo(collectionId, revisionId) {
+    getCollectionGraph(query, collectionId) {
         return __awaiter(this, void 0, void 0, function* () {
             yield this.mQuota.wait();
-            const res = yield this.request(this.mBaseURL + '/collections/{collectionId}/revisions/{revisionId}', this.args({
-                path: this.filter({ collectionId, revisionId }),
-            }));
-            return res.collection_revision;
+            const res = yield this.requestGraph('collection', {
+                id: { type: 'Int', optional: false },
+            }, query, { id: collectionId }, this.args({ path: this.filter({}) }));
+            return res;
         });
     }
-    getCollectionDownloadURLs(collectionId, revisionId, key, expires, gameId) {
+    getCollectionListGraph(query, gameId, count, page) {
         return __awaiter(this, void 0, void 0, function* () {
             yield this.mQuota.wait();
-            let urlPath = '/games/{gameId}/collections/{collectionId}/revisions/{revisionId}/download_link';
-            if ((key !== undefined) && (expires !== undefined)) {
-                urlPath += '?key={key}&expires={expires}';
-            }
-            return this.request(this.mBaseURL + urlPath, this.args({ path: this.filter({ collectionId, revisionId, gameId, key, expires }) }));
+            const res = yield this.requestGraph('collections', {
+                game: { type: 'String', optional: false },
+                count: { type: 'Int', optional: true },
+                page: { type: 'Int', optional: true },
+            }, query, { game: gameId || this.mBaseData.path.game, count, page }, this.args({ path: this.filter({}) }));
+            return res;
+        });
+    }
+    getRevisionGraph(query, revisionId) {
+        return __awaiter(this, void 0, void 0, function* () {
+            yield this.mQuota.wait();
+            const res = yield this.requestGraph('collectionRevision', {
+                id: { type: 'Int', optional: false },
+            }, query, { id: revisionId }, this.args({ path: this.filter({}) }));
+            return res;
         });
     }
     endorseCollection(collectionId, endorseStatus, gameId) {
@@ -559,25 +493,28 @@ class Nexus {
                 if (anonymous) {
                     delete headers['APIKEY'];
                 }
-                const url = anonymous
+                const inputUrl = anonymous
                     ? `${param.API_URL}/feedbacks/anonymous`
                     : `${param.API_URL}/feedbacks`;
-                request.post({
-                    headers,
-                    url,
-                    formData,
-                    timeout: 30000,
-                }, (error, response, body) => {
-                    if (error !== null) {
-                        return reject(error);
-                    }
-                    else if (response.statusCode >= 400) {
-                        return reject(new customErrors_1.HTTPError(response.statusCode, response.statusMessage, body));
-                    }
-                    else {
-                        return resolve(JSON.parse(body));
-                    }
+                const req = lib(inputUrl).request(Object.assign(Object.assign({}, url.parse(inputUrl)), { method: 'POST', headers, timeout: 30000 }), (res) => {
+                    res.setEncoding('utf8');
+                    let rawData = '';
+                    res
+                        .on('data', (chunk) => { rawData += chunk; })
+                        .on('error', err => {
+                        return reject(err);
+                    })
+                        .on('end', () => {
+                        if (res.statusCode >= 400) {
+                            return reject(new customErrors_1.HTTPError(res.statusCode, res.statusMessage, rawData));
+                        }
+                        else {
+                            return resolve(JSON.parse(rawData));
+                        }
+                    });
                 });
+                req.write(querystring.stringify(formData));
+                req.end();
             }));
         });
     }
@@ -613,6 +550,75 @@ class Nexus {
                     }
                 }
                 throw err;
+            }
+        });
+    }
+    makeQueryImpl(query, variables, indent) {
+        return Object.keys(query).reduce((prev, key) => {
+            if (query[key] !== false) {
+                prev += (indent + key);
+                if (typeof query[key] !== 'boolean') {
+                    prev = prev +
+                        ' {\n'
+                        + this.makeQueryImpl(query[key], variables, indent + '  ')
+                        + `${indent}}`;
+                }
+                return prev + '\n';
+            }
+            return prev;
+        }, '') + indent;
+    }
+    makeParameters(parameters) {
+        const serParameter = (par) => {
+            return `${par.type}${par.optional ? '' : '!'}`;
+        };
+        return Object.keys(parameters)
+            .map(key => `\$${key}: ${serParameter(parameters[key])}`)
+            .join(', ');
+    }
+    makeFilter(parameters) {
+        return Object.keys(parameters)
+            .map(key => `${key}: \$${key}`)
+            .join(', ');
+    }
+    makeQuery(name, parameters, query, variables) {
+        return `query ${name}(${this.makeParameters(parameters)}) {\n`
+            + `  ${name}(${this.makeFilter(parameters)}) {${this.makeQueryImpl(query, variables, '    ')}`
+            + '  }\n'
+            + '}';
+    }
+    makeMutation(name, parameters, data) {
+        return `mutation ${name}(${this.makeParameters(parameters)}) {\n`
+            + `  ${name}(${this.makeFilter(parameters)}) { collectionId, revisionId, success }`
+            + '}';
+    }
+    requestGraph(root, parameters, query, variables, args) {
+        return __awaiter(this, void 0, void 0, function* () {
+            args.data = {
+                query: this.makeQuery(root, parameters, query, variables),
+                variables,
+            };
+            const res = yield this.request(this.mGraphBaseURL, args, 'POST');
+            if (res.data) {
+                return res.data[root];
+            }
+            else {
+                throw new Error(res.errors.map(err => err.message).join(', '));
+            }
+        });
+    }
+    mutateGraph(name, parameters, data, args) {
+        return __awaiter(this, void 0, void 0, function* () {
+            args.data = {
+                query: this.makeMutation(name, parameters, data),
+                variables: data,
+            };
+            const res = yield this.request(this.mGraphBaseURL, args, 'POST');
+            if (res.data) {
+                return res.data[name];
+            }
+            else {
+                throw new Error(res.errors.map(err => err.message).join(', '));
             }
         });
     }
